cloud_functions/load/requirements.txt
pandas
google-cloud-bigquery
pandas-gbq
gcsfs
pyarrow 